+++ Editor/HotspotCameraEditor.cs
@@ -1,0 +1,13 @@
+﻿using UnityEditor;
+
+namespace Treasured.UnitySdk
+{
+    [CustomEditor(typeof(HotspotCamera))]
+    internal class HotspotCameraEditor : UnityEditor.Editor
+    {
+        public override void OnInspectorGUI()
+        {
+            // No GUI
+        }
+    }
+}