+++ Editor/Json/JsonValidator.cs
@@ -4,13 +4,17 @@
     {
         public static void ValidateMap(TreasuredMap map)
         {
+            if (string.IsNullOrWhiteSpace(map.Author))
+            {
+                throw new ContextException("Missing Field", "The author field is missing.", map);
+            }
             if (string.IsNullOrWhiteSpace(map.Title))
             {
-                throw new System.MissingFieldException("The title field is missing.");
+                throw new ContextException("Missing Field", "The title field is missing.", map);
             }
             if (string.IsNullOrWhiteSpace(map.Description))
             {
-                throw new System.MissingFieldException("The description field is missing.");
+                throw new ContextException("Missing Field", "The description field is missing.", map);
             }
             foreach (var obj in map.GetComponentsInChildren<TreasuredObject>())
             {
@@ -22,13 +26,40 @@
         {
             if (obj is Hotspot hotspot && hotspot.Camera == null)
             {
-                throw new TargetNotAssignedException($"Camera Transform is not assigned for {obj.name}.", obj);
+                throw new ContextException("Missing reference", $"Camera Transform is not assigned for {obj.name}.", obj);
             }
-            foreach (var action in obj.OnSelected)
+            foreach (var group in obj.OnClick)
             {
-                if (action is SelectObjectAction soa && (soa.Target == null || (soa.Target != null && !soa.Target.gameObject.activeSelf)))
+                foreach (var action in group.Actions)
                 {
-                    throw new TargetNotAssignedException($"The target for Select-Object action is inactive OR is not assigned for {obj.name}.", obj);
+                    if (action is SelectObjectAction soa)
+                    {
+                        if (soa.Target == null || (soa.Target != null && !soa.Target.gameObject.activeSelf))
+                        {
+                            throw new ContextException("Missing reference", $"The target for Select-Object action is inactive OR is not assigned for {obj.name}.", obj);
+                        }
+                        else if(soa.Target.GetComponentInParent<TreasuredMap>() != obj.GetComponentInParent<TreasuredMap>())
+                        {
+                            throw new ContextException("Invalid reference", $"The target set for Select-Object action does not belong to the same map.", obj);
+                        }
+                    }
+                }
+            }
+            foreach (var group in obj.OnHover)
+            {
+                foreach (var action in group.Actions)
+                {
+                    if (action is SelectObjectAction soa)
+                    {
+                        if (soa.Target == null || (soa.Target != null && !soa.Target.gameObject.activeSelf))
+                        {
+                            throw new ContextException("Missing reference", $"The target for OnHover-Object action is inactive OR is not assigned for {obj.name}.", obj);
+                        }
+                        else if (soa.Target.GetComponentInParent<TreasuredMap>() != obj.GetComponentInParent<TreasuredMap>())
+                        {
+                            throw new ContextException("Invalid reference", $"The target set for OnHover-Object action does not belong to the same map.", obj);
+                        }
+                    }
                 }
             }
         }