+++ Runtime/Attributes/PresetAttribute.cs
@@ -2,11 +2,11 @@
 
 namespace Treasured.UnitySdk
 {
-    public class StringSelectorAttribute : PropertyAttribute
+    public class PresetAttribute : PropertyAttribute
     {
         public string[] Values { get; set; } = new string[] { string.Empty };
 
-        public StringSelectorAttribute(params string[] values)
+        public PresetAttribute(params string[] values)
         {
             Values = values;
         }