+++ Runtime/Component/Hitbox.cs
@@ -2,6 +2,7 @@
 
 namespace Treasured.UnitySdk
 {
+    [RequireComponent(typeof(BoxCollider))]
     public sealed class Hitbox : MonoBehaviour
     {
 #if UNITY_EDITOR
@@ -11,7 +12,15 @@ namespace Treasured.UnitySdk
             Matrix4x4 tempMatrix = Gizmos.matrix;
             Gizmos.matrix = Matrix4x4.TRS(this.transform.position, this.transform.rotation, Vector3.one);
 
-            Gizmos.color = TreasuredSDKSettings.Instance ? TreasuredSDKSettings.Instance.hitboxColor : TreasuredSDKSettings.defaultHitboxColor;
+            if (transform.parent != null && transform.parent.GetComponent<Hotspot>())
+            {
+                Gizmos.color = Color.red;
+            }
+            else
+            {
+                Gizmos.color = TreasuredSDKSettings.Instance ? TreasuredSDKSettings.Instance.hitboxColor : TreasuredSDKSettings.defaultHitboxColor;
+            }
+
             Gizmos.DrawCube(Vector3.zero, transform.localScale);
 
             Gizmos.color = tempColor;