+++ Runtime/Data/Actions/Embed/EmbedAction.cs
@@ -1,19 +1,12 @@
-﻿using Newtonsoft.Json;
-using UnityEngine;
+﻿using UnityEngine;
 
 namespace Treasured.UnitySdk
 {
-    [Category("Media")]
+    [Category("Embed")]
     public abstract class EmbedAction : ActionBase
     {
         [SerializeField]
-        [TextArea(3, 3)]
-        private string _src;
-
-        [SerializeField]
         private EmbedPosition _position = EmbedPosition.TopRight;
-
-        public string Src { get => _src; set => _src = value; }
 
         public EmbedPosition Position { get => _position; set => _position = value; }
     }